requirements.txt
asgiref
attrs
Django
django-filter
djangorestframework
drf-spectacular
drf-yasg
inflection
jsonschema
jsonschema-specifications
packaging
pytz
PyYAML
referencing
rpds-py
sqlparse
typing_extensions


